backend/requirements.txt
# =============================================================================
# backend/requirements.txt
# Install with: pip install -r requirements.txt
#
# All packages are chosen for CPU-only operation (Raspberry Pi compatible).
# =============================================================================

# ---- Web framework ----
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0

# ---- Face detection (MediaPipe - replaces MTCNN+TensorFlow, much faster on Pi) ----
# mediapipe has official aarch64 wheels: pip install mediapipe
# NOTE: mtcnn + tensorflow-cpu are removed; MediaPipe handles detection instead.
mediapipe>=0.10.0

# ---- Face embedding (FaceNet via facenet-pytorch) ----
facenet-pytorch>=2.5.3
torch>=2.1.0                 # CPU-only; install separately on RPi:
                             #   pip install torch --index-url https://download.pytorch.org/whl/cpu
torchvision>=0.16.0
Pillow>=10.0.0

# ---- Head pose / liveness (OpenCV) ----
opencv-python-headless>=4.8.0  # headless = no GUI dependency (RPi/server safe)
numpy>=1.24.0

# ---- Recognition (SVM) ----
scikit-learn>=1.3.0

# ---- Evaluation / visualisation ----
matplotlib>=3.8.0
seaborn>=0.13.0              # optional  nicer default themes for eval plots

# ---- Utilities ----
gunicorn>=21.0.0             # production WSGI server

# ---- Raspberry Pi hardware (install only on Pi) ----
# RPi.GPIO is pre-installed on Raspberry Pi OS; uncomment if needed:
# RPi.GPIO>=0.7.1
smbus2>=0.4.3                # I2C for LCD (16x2 HD44780 with PCF8574 backpack)

